+++ src/main/java/gg/archipelago/APClient/network/LocationChecks.java
@@ -9,7 +9,7 @@ import java.util.Set;
 public class LocationChecks extends APPacket {
 
     @SerializedName("locations")
-    public Set<Integer> locations = new HashSet<>();
+    public Set<Long> locations = new HashSet<>();
 
     public LocationChecks() {
         this.cmd = APPacketType.LocationChecks;
