+++ src/main/java/gg/archipelago/APClient/network/LocationScouts.java
@@ -5,9 +5,9 @@ import java.util.ArrayList;
 
 public class LocationScouts extends APPacket {
 
-    public ArrayList<Integer> locations = new ArrayList<>();
+    public ArrayList<Long> locations = new ArrayList<>();
 
-    public LocationScouts(ArrayList<Integer> locations) {
+    public LocationScouts(ArrayList<Long> locations) {
         this.cmd = APPacketType.LocationScouts;
         this.locations = locations;
     }
